+++ ceammc/ext/src/array/array_fill.h
@@ -26,7 +26,13 @@ public:
     void onFloat(float f);
     void onList(const AtomList& l);
 
+    void m_gauss(t_symbol* m, const AtomList& l);
+    void m_uniform(t_symbol* m, const AtomList& l);
     void m_fill(t_symbol* m, const AtomList& l);
+    void m_sin(t_symbol* m, const AtomList& l);
+    void m_pulse(t_symbol* m, const AtomList& l);
+    void m_saw(t_symbol* m, const AtomList& l);
+    void m_tri(t_symbol* m, const AtomList& l);
     AtomList parseRange(const AtomList& args, size_t* from, size_t* to) const;
 
 private: