+++ ceammc/ext/src/array/mod_array.cpp
@@ -1,7 +1,14 @@
 #include "mod_array.h"
 #include "array_bpm.h"
+#include "array_hist.h"
+#include "array_mean.h"
+#include "array_rms.h"
 #include "array_set.h"
+#include "array_stddev.h"
 #include "array_stretch.h"
+#include "array_sum.h"
+#include "array_sum2.h"
+#include "array_variance.h"
 #include "array_vline_play.h"
 #include "array_window.h"
 
@@ -25,6 +32,13 @@ void ceammc_array_setup()
     setup_array0x2es2p();
     setup_array0x2eset();
     setup_array0x2estretch();
+    setup_array_hist();
+    setup_array_mean();
+    setup_array_rms();
+    setup_array_stddev();
+    setup_array_sum();
+    setup_array_sum2();
+    setup_array_variance();
     setup_array_vline_play();
     setup_array_window();
 }