+++ ceammc/CHANGELOG.md
@@ -8,6 +8,21 @@
   - foldable categories added
 - outer/inner label added for almost all UI widgets
 - ui.matrix read/write methods added to save/load in text format
+- array.hist added
+- array.mean added
+- array.rms added
+- array.stddev added
+- array.sum added
+- array.sum2 added
+- array.variance added
+- hw.apple_smc (system management control) added
+- hw.apple_sms (sudden motion sensor) added
+- hw.cpu_temp added
+- hw.display added
+- hw.kdb_light added
+- array.fill sin method added
+- array.fill gauss method added
+- array.fill uniform method added
 
 ## [0.7]
 ### Added: